+++ src/app/components/home/home.component.ts
@@ -84,7 +84,7 @@ export class HomeComponent implements OnInit {
   hideMenuOptions() {
     this.g.wdLog(['hideMenuOptions']);
     // this.g.isOpenMenuOptions = false;
-    this.g.setParameter('isOpenMenuOptions', false);
+    this.g.setParameter('isOpenMenuOptions', false, true);
   }